requirements.txt
# GovScan 필수 의존성 패키지

# 핵심 라이브러리
pandas>=1.5.0
openpyxl>=3.1.0
packaging>=21.0
Jinja2>=3.1.0

# 네트워크 스캔
python-nmap>=0.7.1

# 웹 애플리케이션
Flask>=2.3.0
Werkzeug>=2.3.0

# 데이터베이스 연결 (선택적)
psycopg2-binary>=2.9.0

# SSH 연결 (선택적)
paramiko>=3.0.0

# MySQL 연결 (선택적)  
mysql-connector-python>=8.0.0

# 기타 유틸리티
requests>=2.28.0
PyYAML>=6.0.0
python-dateutil>=2.8.0

# 개발/테스트 도구 (선택적)
pytest>=7.0.0
pytest-cov>=4.0.0

# 보안 관련
cryptography>=3.4.8

# 파일 처리
et-xmlfile>=2.0.0

# 네트워크 분석
maxminddb>=2.7.0
geoip2>=5.1.0

# HTTP 클라이언트
aiohttp>=3.8.0
aiosignal>=1.3.0
attrs>=22.0.0
certifi>=2022.0.0
charset-normalizer>=3.0.0
frozenlist>=1.3.0
idna>=3.4
multidict>=6.0.0
yarl>=1.8.0

# 시스템 유틸리티
six>=1.16.0
tzdata>=2023.3
urllib3>=1.26.0

# GUI (선택적)
PySide6>=6.5.0